(rename) : 패키지 명 변경 #EAT-20 코드 컨벤션에 맞게 패키지명에 대문제 -> 소문자로 변경
